requirements.txt
flask
flask-cors
psycopg2-binary>=2.9.9
git+https://github.com/openai/swarm.git
python-dotenv>=0.19.0